Rework dose normalization. Separate model scoring.
* Change how dose normalization works (e.g. skull-base study). Instead
of normalizing to the mean target dose, we now normalize to the
prescribed dose value found in the Patient config file. This means that
we can now predict a DVH without needing a plan to already exist to get
the correct normalization. Previously, the predicted DVH had to be
manually normalized to a realistic dose scale.

* Move model scoring into separate function. This is compatible with
scikit-learn cross-validation, etc.

* Rename ShellFieldModel to CoplanarShellModel